moved `css` plugin to its own package
